fix: split code into modules
